Fixed compiler warnings

+++ src/lib/SCDevice2Web/SCDevice2Web/SCDWSettings.h
@@ -1,4 +1,4 @@
 #import <Foundation/Foundation.h>
 
-extern NSUInteger JDWFileSystemURLPort;
+extern long JDWFileSystemURLPort;
 extern NSString* const JDWFileSystemURLScheme;
