+++ src/lib/SCDevice2Web/SCDevice2Web/SCDWSettings.m
@@ -1,4 +1,4 @@
 #import "SCDWSettings.h"
 
-NSUInteger JDWFileSystemURLPort = 33445;
+long JDWFileSystemURLPort = 33445;
 NSString* const JDWFileSystemURLScheme = @"scmobile";
